comment delete confirm working on backend. Added backend confirmation that comment.user == current_user

+++ app/controllers/comments_controller.rb
@@ -13,10 +13,15 @@ class CommentsController < ApplicationController
 
 
   def destroy
-    comment = Comment.find(params[:id])
-    post = comment.post
-    comment.destroy
-    render json: post
+    comment = Comment.find_by(id: params[:id], user_id: @current_user[:id])
+    if comment
+      post = comment.post
+      comment.destroy
+      render json: post
+    else
+      render json: { errors: comment.errors.full_messages }, status: :unprocessable_entity
+    end
+
   end
 
   private
